# apps/frontend/openapi.yaml
openapi: 3.1.0
info:
  title: ALL-Flow API
  version: 1.0.0
  description: |
    ALL-Flow 사내 협업 시스템 백엔드 API 스펙.
    프론트엔드 `lib/types.ts` 와 양방향으로 동기화됩니다.

    인증: Bearer JWT (next-auth v5 발급) — 모든 보호된 엔드포인트에 필요.
servers:
  - url: https://api.allflow.omelet.ai/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Local dev
security:
  - bearerAuth: []
...
paths:
...
  /projects:
    get:
      tags: [projects]
      summary: 프로젝트 목록
      parameters:
        - { name: status, in: query, schema: { $ref: '#/components/schemas/StatusKey' } }
        - { name: q, in: query, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Project' }
    post:
      tags: [projects]
      summary: 프로젝트 생성
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
  /projects/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [projects]
      summary: 프로젝트 단건 조회
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
    patch:
      tags: [projects]
      summary: 프로젝트 수정
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectPatch' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
...
  /issues:
    get:
      tags: [issues]
      summary: 이슈 목록
      parameters:
        - { name: status, in: query, schema: { $ref: '#/components/schemas/IssueStatus' } }
        - { name: prio, in: query, schema: { $ref: '#/components/schemas/IssuePrio' } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Issue' }
    post:
      tags: [issues]
      summary: 이슈 생성
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IssueCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Issue' }
...
  /issues/{id}/transition:
    post:
      tags: [issues]
      summary: 이슈 상태 전이
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { $ref: '#/components/schemas/IssueStatus' }
                comment: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Issue' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    StatusKey:
      type: string
      enum: [todo, doing, review, done, blocked]
...
    Project:
      type: object
      required: [id, name, code, color, progress, status, due, members, tasks]
      properties:
        id: { type: string }
        name: { type: string }
        code: { type: string }
        color: { type: string }
        progress: { type: integer, minimum: 0, maximum: 100 }
        status: { $ref: '#/components/schemas/StatusKey' }
        due: { type: string, format: date }
        members:
          type: array
          items: { type: string }
        tasks:
          type: object
          required: [total, done]
          properties:
            total: { type: integer }
            done: { type: integer }
    ProjectCreate:
      type: object
      required: [name, code]
      properties:
        name: { type: string }
        code: { type: string }
        color: { type: string }
        due: { type: string, format: date }
    ProjectPatch:
      type: object
      properties:
        name: { type: string }
        progress: { type: integer }
        status: { $ref: '#/components/schemas/StatusKey' }
        due: { type: string, format: date }
...
    IssueSev:
      type: string
      enum: [critical, high, med, low]
    IssuePrio:
      type: string
      enum: [P0, P1, P2, P3]
    IssueStatus:
      type: string
      enum: [open, in-progress, in-review, resolved]
    Issue:
      type: object
      required:
        [id, title, proj, projColor, sev, prio, status, assignee, reporter,
         tags, created, sla, slaPct, comments, linked]
      properties:
        id: { type: string }
        title: { type: string }
        proj: { type: string }
        projColor: { type: string }
        sev: { $ref: '#/components/schemas/IssueSev' }
        prio: { $ref: '#/components/schemas/IssuePrio' }
        status: { $ref: '#/components/schemas/IssueStatus' }
        assignee: { type: string }
        reporter: { type: string }
        tags:
          type: array
          items: { type: string }
        created: { type: string }
        sla: { type: string }
        slaPct: { type: integer }
        comments: { type: integer }
        linked: { type: integer }
        resolved: { type: boolean }
...
    IssueCreate:
      type: object
      required: [title, proj, assignee, reporter, sev, prio]
      properties:
        title: { type: string }
        proj: { type: string }
        assignee: { type: string }
        reporter: { type: string }
        sev: { $ref: '#/components/schemas/IssueSev' }
        prio: { $ref: '#/components/schemas/IssuePrio' }
        tags: { type: array, items: { type: string } }
...